hide warnings about readl/writel

+++ drivers/video/msm/mdp_debugfs.c
@@ -661,7 +661,7 @@ static int mddi_reg_read(int ndx)
 
 	mdp_pipe_ctrl(MDP_CMD_BLOCK, MDP_BLOCK_POWER_ON, FALSE);
 	while (reg->name) {
-		data = readl((u32)base + reg->off);
+		data = readl(base + reg->off);
 		len = snprintf(bp, dlen, "%s:0x%08x\t\t= 0x%08x\n",
 					reg->name, reg->off, data);
 		tot += len;
@@ -1014,7 +1014,7 @@ static ssize_t dbg_reg_read(
 			if (num >= dbg_count)
 				break;
 		}
-		data = readl((u32)cp + off);
+		data = readl(cp + off);
 		*bp++ = '\n';
 		--dlen;
 		tot++;
